+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/backend/engine/checkers/PackageChecker.kt
@@ -1,5 +1,7 @@
 package com.github.warningimhack3r.npmupdatedependencies.backend.engine.checkers
 
+import com.github.warningimhack3r.npmupdatedependencies.settings.NUDSettingsState
+
 abstract class PackageChecker {
     companion object {
         private val ONLY_DIGIT_X_OR_DOT = Regex("^[\\dx.]+$")
@@ -45,7 +47,7 @@ abstract class PackageChecker {
                 // Anything that is not supported is not upgradable
                 !isComparatorSupported(this) -> false
                 // If the version is "static", upgradable only if it includes (ends with?) a `x` or `X`
-                ONLY_DIGIT_X_OR_DOT.matches(this) -> lowercase().contains("x")
+                ONLY_DIGIT_X_OR_DOT.matches(this) -> lowercase().contains("x") || NUDSettingsState.instance.checkStaticComparators
                 // Symbols prefixes
                 startsWith("<") || startsWith("<=") -> false
                 // Anything else is considered upgradable by default